+++ docs/conf.py
@@ -16,14 +16,6 @@ import sys
 import os
 import re
 
-# Imports used by code for autoautosummary
-import inspect
-from sphinx.ext.autosummary import Autosummary
-from sphinx.ext.autosummary import get_documenter
-from docutils.parsers.rst import directives
-from sphinx.util.inspect import safe_getattr
-from sphinx.util import logging
-
 from pywbem._cim_constants import DEFAULT_NAMESPACE
 
 
@@ -78,6 +70,7 @@ extensions = [
     # but since it already provides a full TOC in the navigation pane, the
     # sphinxcontrib.fulltoc extension is not needed.
     'sphinx_rtd_theme',
+    'autodocsumm',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -430,7 +423,7 @@ autoclass_content = "both"
 # Selects if automatically documented members are sorted alphabetically
 # (value 'alphabetical'), by member type (value 'groupwise') or by source
 # order (value 'bysource'). The default is alphabetical.
-autodoc_member_order = "bysource"
+autodoc_member_order = "alphabetical"
 
 # This value is a list of autodoc directive options (flags) that should be
 # automatically applied to all autodoc directives. The supported options
@@ -464,6 +457,15 @@ autodoc_docstring_signature = True
 # process.
 autodoc_mock_imports = []
 
+# -- Options for autodocsumm extension ------------------------------------
+# For documentation, see
+# https://autodocsumm.readthedocs.io/en/latest/
+
+# Default options to be applied by autodocsumm to ::autoclass and all
+# other ::auto... directives. While it would nicely allow specifying autoclass
+# options that have values, autodocsumm applies this setting to all directives,
+# and autosummary cannot be turned off by directive, so we don't use it.
+autodoc_default_options = {}
 
 # -- Options for intersphinx extension ------------------------------------
 # For documentation, see
@@ -534,164 +536,3 @@ extlinks = {
   'nbview': ('https://nbviewer.jupyter.org/github/pywbem/pywbem/blob/master/docs/notebooks/%s', ''),
   'nbdown': ('https://github.com/pywbem/pywbem/raw/master/docs/notebooks/%s', '')
 }
-
-# -- Support for autoautosummary ----------------------------------------------
-#
-# Idea taken from https://stackoverflow.com/a/30783465/1424462
-#
-
-class AutoAutoSummary(Autosummary):
-    """
-    Sphinx extension that automatically generates a table of public methods or
-    attributes of a class, using the AutoSummary extension.
-    (i.e. each row in the table shows the method or attribute name with a
-    link to the full description, and a one-line brief description).
-
-    Usage in RST source::
-
-        .. autoclass:: path.to.class
-           :<autoclass-options>:
-
-           .. rubric:: Methods
-
-           .. autoautosummary:: path.to.class
-              :methods:
-
-           .. rubric:: Attributes
-
-           .. autoautosummary:: path.to.class
-              :attributes:
-
-           .. rubric:: Details
-
-    """
-
-    option_spec = {
-        'methods': directives.unchanged,
-        'attributes': directives.unchanged
-    }
-    option_spec.update(Autosummary.option_spec)
-
-    required_arguments = 1  # Fully qualified class name
-
-    def __init__(self, *args, **kwargs):
-        self._logger = logging.getLogger(__name__)  # requires Sphinx 1.6.1
-        self._log_prefix = "conf.py/AutoAutoSummary"
-        self._excluded_classes = ['BaseException']
-        super(AutoAutoSummary, self).__init__(*args, **kwargs)
-
-    def _get_members(self, class_obj, member_type, include_in_public=None):
-        """
-        Return class members of the specified type.
-
-        class_obj: Class object.
-
-        member_type: Member type ('method' or 'attribute').
-
-        include_in_public: set/list/tuple with member names that should be
-          included in public members in addition to the public names (those
-          starting without underscore).
-
-        Returns:
-          tuple(public_members, all_members): Names of the class members of
-            the specified member type (public / all).
-        """
-        try:
-            app = self.state.document.settings.env.app
-        except AttributeError:
-            app = None
-        if not include_in_public:
-            include_in_public = []
-        all_members = []
-        for member_name in dir(class_obj):
-            try:
-                documenter = get_documenter(
-                    app,
-                    safe_getattr(class_obj, member_name),
-                    class_obj)
-            except AttributeError:
-                continue
-            if documenter.objtype == member_type:
-                all_members.append(member_name)
-        public_members = [x for x in all_members
-                  if x in include_in_public or not x.startswith('_')]
-        return public_members, all_members
-
-    def _get_def_class(self, class_obj, member_name):
-        """
-        Return the class object in MRO order that defines a member.
-
-        class_obj: Class object that exposes (but not necessarily defines) the
-          member. I.e. starting point of the search.
-
-        member_name: Name of the member (method or attribute).
-
-        Returns:
-          Class object that defines the member.
-        """
-        member_obj = getattr(class_obj, member_name)
-        for def_class_obj in inspect.getmro(class_obj):
-            if member_name in def_class_obj.__dict__:
-                if def_class_obj.__name__ in self._excluded_classes:
-                    return class_obj  # Fall back to input class
-                return def_class_obj
-        self._logger.warning(
-            "%s: Definition class not found for member %s.%s, "
-            "defaulting to class %s",
-            self._log_prefix, class_obj.__name__, member_name,
-            class_obj.__name__)
-        return class_obj  # Input class is better than nothing
-
-    def run(self):
-
-        try:
-            full_class_name = str(self.arguments[0])
-            module_name, class_name = full_class_name.rsplit('.', 1)
-            module_obj = __import__(module_name, globals(), locals(),
-                                    [class_name])
-            class_obj = getattr(module_obj, class_name)
-            if 'methods' in self.options:
-                _, methods = self._get_members(
-                    class_obj, 'method', ['__init__'])
-                self.content = []
-                for method in methods:
-                    if method.startswith('_'):
-                        # Skip private methods
-                        continue
-                    def_class = self._get_def_class(class_obj, method)
-                    def_module_name = def_class.__module__
-                    if def_module_name.startswith('pywbem'):
-                        def_module_name = def_module_name.split('.')[0]
-                    content_str = "~%s.%s.%s" % (
-                        def_module_name,
-                        def_class.__name__,
-                        method)
-                    self.content.append(content_str)
-            elif 'attributes' in self.options:
-                _, attributes = self._get_members(class_obj, 'attribute')
-                self.content = []
-                for attrib in attributes:
-                    if attrib.startswith('_'):
-                        # Skip private attributes
-                        continue
-                    def_class = self._get_def_class(class_obj, attrib)
-                    def_module_name = def_class.__module__
-                    if def_module_name.startswith('pywbem'):
-                        def_module_name = def_module_name.split('.')[0]
-                    content_str = "~%s.%s.%s" % (
-                        def_module_name,
-                        def_class.__name__,
-                        attrib)
-                    self.content.append(content_str)
-
-        except Exception as exc:
-            self._logger.error(
-                "%s: Internal error: %s: %s",
-                self._log_prefix, exc.__class__.__name__, exc)
-
-        finally:
-            return super(AutoAutoSummary, self).run()
-
-
-def setup(app):
-    app.add_directive('autoautosummary', AutoAutoSummary)